Update manage-payg-transition.ps1
Added parameter SkipDownload with default to $false, which allows customers to manually download the modify-azure-sql-license-type.ps1 and the modify-arc-sql-license-type.ps1 in situations where  scripts are not allowed to download files from the internet. 

if user opt for manually download the 2 .ps1 files, these files need to be placed under a subfolder named PayTransitionDownloads where this main script is running from. 

During the execution, the script will validate the 2 .ps1 files are present

last but not least, if the user uses the SkipDownload parameter with value $True, that subfolder is not deleted after execution of the script